Move printer utils into printer package
