+++ pkg/rakkess/rakkess.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/corneliusweig/rakkess/pkg/rakkess/client"
 	"github.com/corneliusweig/rakkess/pkg/rakkess/options"
+	"github.com/corneliusweig/rakkess/pkg/rakkess/printer"
 	"github.com/corneliusweig/rakkess/pkg/rakkess/util"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
@@ -52,7 +53,7 @@ func Rakkess(ctx context.Context, opts *options.RakkessOptions) error {
 		return errors.Wrap(err, "check resource access")
 	}
 
-	util.PrintResults(opts.Streams.Out, opts.Verbs, outputFormat(opts), results)
+	printer.PrintResults(opts.Streams.Out, opts.Verbs, outputFormat(opts), results)
 
 	if namespace == nil || *namespace == "" {
 		fmt.Fprintf(opts.Streams.Out, "No namespace given, this implies cluster scope (try -n if this is not intended)\n")
@@ -61,9 +62,9 @@ func Rakkess(ctx context.Context, opts *options.RakkessOptions) error {
 	return nil
 }
 
-func outputFormat(o *options.RakkessOptions) util.OutputFormat {
+func outputFormat(o *options.RakkessOptions) printer.OutputFormat {
 	if o.Output == "ascii-table" {
-		return util.ASCIITable
+		return printer.ASCIITable
 	}
-	return util.IconTable
+	return printer.IconTable
 }
